pyproject.toml
[project]
name = "cli-lite"
version = "0.10.0"
description = "A simple framework to build a cli tool, base on Alconna"
authors = [
    {name = "RF-Tar-Railt",email = "rf_tar_railt@qq.com"}
]
dependencies = [
    "arclet-alconna>=1.8.19",
    "arclet-alconna-tools>=0.7.7",
    "importlib-metadata>=8.2.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
keywords = ["commandline", "cli-tool", "plugin", "cli", "litecli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
[project.urls]
Homepage = "https://github.com/RF-Tar-Railt/cli-lite"
"Bug Reports" = "https://github.com/RF-Tar-Railt/cli-lite/issues"
Source = "https://github.com/RF-Tar-Railt/cli-lite"

[tool.pdm]

[tool.pdm.build]
includes = ["clilte"]

[tool.pdm.dev-dependencies]
dev = [
    "fix-future-annotations>=0.5.0",
]
[build-system]
requires = ["pdm.backend"]
build-backend = "pdm.backend"
